fix: move getDefaultAppRole to utility file to fix build
Non-async exports in 'use server' files cause Next.js build errors.
Move getDefaultAppRole to lib/utils/roles.ts.
