+++ apps/web/lib/actions/personen.ts
@@ -13,6 +13,7 @@ import type {
   UserRole,
 } from '../supabase/types'
 import { sanitizeSearchQuery } from '../utils/search'
+import { getDefaultAppRole } from '../utils/roles'
 
 const USE_DUMMY_DATA = !process.env.NEXT_PUBLIC_SUPABASE_URL
 
@@ -834,24 +835,6 @@ export async function resendInvitation(
 // =============================================================================
 // Bulk Invite (Issue #327)
 // =============================================================================
-
-/**
- * Map a person's Vereins-Rolle to the default app role
- */
-export function getDefaultAppRole(rolle: Rolle): UserRole {
-  switch (rolle) {
-    case 'vorstand':
-      return 'VORSTAND'
-    case 'mitglied':
-    case 'regie':
-    case 'technik':
-      return 'MITGLIED_AKTIV'
-    case 'gast':
-      return 'MITGLIED_PASSIV'
-    default:
-      return 'MITGLIED_AKTIV'
-  }
-}
 
 export interface BulkInviteResult {
   total: number
